--- features/learn/src/main/kotlin/com/teacherworkout/features/learn/themes/ThemesFragment.kt
@@ -1,17 +1,0 @@
-package com.teacherworkout.features.learn.themes
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
-import com.teacherworkout.features.learn.R
-
-class ThemesFragment : Fragment() {
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        return inflater.inflate(R.layout.fragment_themes, container, false)
-    }
-}
